Update tests and move old to /old/
