+++ aiidalab_optimade/__init__.py
@@ -8,11 +8,12 @@ from json import load
 
 # pylint: disable=unused-import
 from aiidalab_optimade.importer import OptimadeImporter  # noqa
-from aiidalab_optimade.optimade import OptimadeStructureImport  # noqa
 
-path_to_metadata = path.join(path.dirname(__file__), path.pardir, "metadata.json")
-with open(path_to_metadata, "r") as fp:
-    metadata = load(fp)
+# from aiidalab_optimade.optimade import OptimadeStructureImport  # noqa
+
+PATH_TO_METADATA = path.join(path.dirname(__file__), path.pardir, "metadata.json")
+with open(PATH_TO_METADATA, "r") as fp:
+    METADATA = load(fp)
 
 # In order to update version, change it in `metadata.json`
-__version__ = metadata["version"]
+__version__ = METADATA["version"]
