+++ aiidalab_optimade/utils.py
@@ -1,81 +1,60 @@
-import json
-from six.moves.urllib.request import urlopen
+import requests
 from aiidalab_optimade import exceptions as exc
 
 TIMEOUT_SECONDS = 30  # Seconds before timeout is raised
 
-TEMP_PROVIDERS_URL = "https://raw.githubusercontent.com/Materials-Consortia/OPTiMaDe/247aa7bfa7d260719297d5889a2844bdddd5d4bc/providers.json"
+PROVIDERS_URL = "https://www.optimade.org/providers/links"
 
 
-def fetch_providers(providers_url=None):
-    """ Fetch OPTiMaDe database providers
-    Fetch JSON file from GitHub.com/Materials-Consortia/OPTiMaDe/providers.json
+def fetch_providers(providers_url: str = None) -> list:
+    """ Fetch OPTiMaDe database providers (from Materials-Consortia)
+
     :param providers_url: String with URL to providers.json file
     """
     if providers_url and not isinstance(providers_url, str):
         raise TypeError("providers_url must be a string")
 
     if not providers_url:
-        providers_url = TEMP_PROVIDERS_URL
+        providers_url = PROVIDERS_URL
 
     try:
-        providers = urlopen(providers_url, timeout=TIMEOUT_SECONDS)
-        providers = providers.read()
+        providers = requests.get(providers_url, timeout=TIMEOUT_SECONDS)
     except Exception:
         raise exc.NonExistent("The URL cannot be opened: {}".format(providers_url))
+    else:
+        providers = providers.json()
 
-    # Load providers.json
-    providers = json.loads(providers)
-
-    # Get dict of providers
-    providers = providers["data"][0]["attributes"]["providers"]
-
-    return providers
+    # Return list of providers
+    return providers["data"]
 
 
-def validate_provider_details(details):
+def validate_provider_details(details: dict) -> dict:
     """Validate dict of details from providers.json"""
-    from string import capwords
-
     if not details or not isinstance(details, dict):
         raise exc.InputError("Please specify 'details', it must be a dict")
 
-    if details["index_base_url"] is None:
-        details["index_base_url"] = ""
-
-    if details["description"] is None:
-        details["description"] = ""
-
-    details["description"] = capwords(details["description"])
+    for field in details:
+        if details[field] is None:
+            details[field] = ""
 
     return details
 
 
 def get_list_of_database_providers():
     """ Get list of database providers
+
     Return formatted list of tuples to use for a dropdown-widget.
     """
-    # Fetch providers from official providers.json
-    providers_dict = fetch_providers()
-
-    # Initialize
+    providers = fetch_providers()
     res = []
 
-    # Go through dict of providers
-    for provider, details in providers_dict.items():
+    for provider in providers:
         # Skip if "exmpl"
-        if provider == "exmpl":
+        if provider["id"] == "exmpl":
             continue
 
-        # Validate details
-        details = validate_provider_details(details)
-
-        res_dict = {}
-        res_dict["name"] = provider
-        res_dict["index"] = details["index_base_url"]
-
-        provider_name = details["description"]
-
-        res.append((provider_name, res_dict))
+        attributes = provider["attributes"]
+        provider_name = attributes.pop("name")
+        res.append((provider_name, attributes))
 
     return res
